fix: create separate types.py module to avoid import order int __init__.py that fails ruff formatting
