+++ nx_parallel/utils/__init__.py
@@ -1,15 +1,2 @@
-import networkx as nx
-from typing import Union
-from enum import Enum
-
-NX_GTYPES = Union[nx.Graph, nx.DiGraph, nx.MultiGraph, nx.MultiDiGraph]
-
-
-class GraphIteratorType(Enum):
-    NODE = "node"
-    EDGE = "edge"
-    ISOLATE = "isolate"
-
-
 from .chunk import *
 from .decorators import *
